+++ src/main/kotlin/com/asap/asapbackend/domain/user/application/dto/CreateUser.kt
@@ -2,7 +2,7 @@ package com.asap.asapbackend.domain.user.application.dto
 
 import com.asap.asapbackend.domain.child.domain.model.Child
 import com.asap.asapbackend.domain.child.domain.model.Gender
-import com.asap.asapbackend.domain.classroom.domain.model.Grade
+import com.asap.asapbackend.domain.classroom.domain.vo.Grade
 import com.asap.asapbackend.domain.menu.domain.model.Allergy
 import com.asap.asapbackend.domain.user.domain.model.Agreement
 import com.asap.asapbackend.domain.user.domain.model.PhoneNumber
